refactor(settings_view): split shortcut tables + keystroke marshalling into shortcuts.rs
Lifts the entire "shortcuts data + keystroke spec marshalling"
section out of mod.rs into `settings_view::shortcuts`:

  * Public surface (re-exported as `settings_view::*`):
    - `SHORTCUT_ACTIONS` (canonical id list, display order)
    - `effective_shortcut(action, overrides)` (resolves a row's
      live binding, honouring overrides over per-OS defaults)
    - `spec_to_gpui_binding(spec)` (Tauri-style spec →
      gpui keybinding string for `cx.bind_keys`)
  * `pub(super)` helpers used by `SettingsView` itself:
    - `shortcut_label` (display label per action id)
    - `format_spec(&Keystroke)` (Tauri-style spec for persistence)
    - `parse_spec(spec)` (reverse, for rendering the live pill)
    - `any_modifier(&Modifiers)` (recording-state Escape gating)
  * Module-private:
    - `default_for_action` (per-OS — two cfg-gated copies)
    - `canonical_key_for_storage` / `canonical_key_for_display`
      (normalisation between gpui key names and Tauri storage)

Visibility plan keeps the original `pub(crate)` for the three
items main.rs consumes via `settings_view::*`, narrows the four
internal helpers to `pub(super)`, and leaves the helpers used
only inside shortcuts.rs unannotated.

mod.rs drops by ~315 lines (3,763 → 3,452). The
`std::collections::HashMap` + `gpui::{Keystroke, Modifiers}`
imports come out of mod.rs along with the moved code — no
remaining caller named those types.

Verified: `cargo clippy --all-targets -- -D warnings` and
`cargo test` both pass (63 tests, unchanged).
